backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.41
alembic==1.12.1
pydantic==2.11.7
pydantic-settings==2.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==5.4.0
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
email-validator==2.2.0
redis==5.0.1
groq==0.4.1
